+++ tests/suite/extension.test.ts
@@ -1,8 +1,8 @@
 import * as assert from 'assert';
 import * as vscode from 'vscode';
 
-suite('Extension Test Suite', () => {
-	test('Should start extension vonage.vscode', async () => {
+suite('Extension', () => {
+	test('should start extension vonage.vscode', async () => {
 		const started = vscode.extensions.getExtension('vonage.vscode');
 		assert.notStrictEqual(started, undefined);
 		if (started) {