+++ tests/suite/prompts/survey.test.ts
@@ -1,0 +1,63 @@
+import Sinon from 'sinon';
+import chai from 'chai';
+import moment from 'moment';
+import { SurveyPrompt } from '../../../src/prompts';
+import { StorageKeys } from '../../../src/enums';
+import { TestMemento } from '../../mocks';
+
+chai.should();
+
+suite('Prompt:Survey', function () {
+
+  let storage: TestMemento;
+  let surveyPrompt: SurveyPrompt;
+
+  this.beforeEach(function() {
+    storage = new TestMemento();
+    storage.storage = new Map();
+    surveyPrompt = new SurveyPrompt(storage);
+  });
+
+  test(`should not show if user selected to never show again`, async function () {
+    await storage.update(StorageKeys.doNotShowSurveyPromptAgain, true);
+
+    const shouldShow = surveyPrompt.shouldShowBanner();
+
+    shouldShow.should.eq(false);
+  });
+
+  test(`should not show if user has seen message within past 12 weeks`, async function () {
+    const currentEpoch = moment().valueOf();
+    await storage.update(StorageKeys.lastSurveyDate, currentEpoch);
+
+    const shouldShow = surveyPrompt.shouldShowBanner();
+    shouldShow.should.eq(false);
+
+  });
+
+  test(`should not show if not in 20% sampling`, function () {
+    const getRandomIntStub = Sinon
+      .stub(surveyPrompt, 'getRandomInt')
+      .callsFake((max: number): number => {
+        return 80;
+      });
+
+    const shouldShow = surveyPrompt.shouldShowBanner();
+    shouldShow.should.eq(false);
+    getRandomIntStub.calledOnce.should.be.true;
+    getRandomIntStub.restore();
+  });
+
+  test(`should show if in 20% sampling`, function () {
+    const getRandomIntStub = Sinon
+      .stub(surveyPrompt, 'getRandomInt')
+      .callsFake((max: number): number => {
+        return 10;
+      });
+
+    const shouldShow = surveyPrompt.shouldShowBanner();
+    shouldShow.should.eq(true);
+    getRandomIntStub.calledOnce.should.be.true;
+    getRandomIntStub.restore();
+  });
+});